+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.9.0](https://github.com/Kong/volcano-sdk-ruby/compare/v0.8.0...v0.9.0) (2026-09-18)
+
+
+### Features
+
+* **locks:** add recoverable acquisition with caller-owned tokens ([#140](https://github.com/Kong/volcano-sdk-ruby/issues/140)) ([794ac9b](https://github.com/Kong/volcano-sdk-ruby/commit/794ac9bddda42fe94e0b10cbc57d54d919a31cdd))
+
+## [0.8.0](https://github.com/Kong/volcano-sdk-ruby/compare/v0.7.2...v0.8.0) (2026-09-18)
+
+
+### Features
+
+* **auth:** add token bootstrap with server session continuity ([#138](https://github.com/Kong/volcano-sdk-ruby/issues/138)) ([6fb69b5](https://github.com/Kong/volcano-sdk-ruby/commit/6fb69b5f54f09444c4a520663f27c16ab5603752))
+
 ## [0.7.2](https://github.com/Kong/volcano-sdk-ruby/compare/v0.7.1...v0.7.2) (2026-09-17)